Pass self.cacheDir to DBD() from MultiDBD()
MultiDBD() calls __update_dbd_inventory() which then calls DBD() for each filename. In this call, the cacheDir is not passed and therefore DBD's self.cacheDir is set to the default, even if MultiDBD has a specified cacheDir. I have now passed self.cacheDir in the call to DBD() within __update_dbd_inventory().