Fixes issue #5: not throwing an exception when cache file is missing.
If a DBD() instance is created, a missing cache file throws an error.
If a MultiDBD() instance is created, all files are read. Only those that match
the criteria are checked for an existing cache file. If not all files have
access to a cache file, the exception is thrown afterall.