Bug fixes:
All data points with a value inf would previously be dropped, causing
a time offset for valid data after this occurrence. Bug reported by
Rob Cermak.

The .get() method of dbdreader is meant to get one or more variables in
the form of two arrays packed in a tuple: (time, value). The time
paramater (for example m_present_time) is always read implicitly. When
calling .get("m_present_time") undefined behaviour occurred. This is now
fixed and a tuple of two identical time vectors are returned.