Completes implementation of reading compressed data files
Implements decompression algorithms in the C-extension
Some fixes in dbdreader.py
If cache file is missing, now it is tried to see if a ccc file is
present, decoded it and try again.

Approach for linux:
   if data are to be read from a compressed file, then the compressed
file is decompressed and written to a file in memory; all other routines
to extract the data will work on this in-memory file. A decompressed
file will NOT be created.

Approach for windows:
   Because windows lacks a default implementation of fmemopen, the linux
approach cannot be followed. There are some third party solutions on
github, but for now I chose an easier way: if a compressed file is read,
it is first decompressed, and then written to file using the normal
extension. Then this file is processed. It is assumed that the directory
where the compressed files are located, is writeable.

This version is not test on windows yet.