Makes MultiDBD's get_CTD_sync compatible with RBR CTDs
get_CTD_sync method now tests for ctd timestamp variables that are
specific for Seabird and RBR CTDs. To prevent all files to be read
to be traversed completed, the get() methods of DBD and MultiDBD
allow for single parameter requests, to limit the number of values
to be read.