Fixes missing parameter /w MulitDBD's get if present in at least 1 file
If a parameter is missing in one or more files, but present in at least
one, when calling MultiDBD.get or MultiDBD.get_sync, then no error is
thrown. Nans are inserted if interpolation is required, otherwise, only
data are returned as far as data exist. Two new tests have been added to
test this behaviour.