Introduces interpolation functions in user land
For get_sync() the user can specify an interpolation function factory
that will be used instead of the default to interpolate the parameters
supplied to get_sync(). The default behaviour has not changed, but it is
now possible to write your own interpolating function, as well as
writing interpolation functions specific to a parameter.